+++ src/components/VvTab/VvTab.vue
@@ -1,53 +1,77 @@
 <script setup lang="ts">
 	import { VvTabProps, VvTabEvents } from '@/components/VvTab'
-	import type { NavItem } from '@/components/VvNav'
 	import VvNav from '@/components/VvNav/VvNav.vue'
+	import type { NavItemTab } from '@/types/nav'
 
 	const props = defineProps(VvTabProps)
 	const emit = defineEmits(VvTabEvents)
 	const { modifiers, items } = toRefs(props)
-	const activeTab: Ref<string | null> = ref(null)
+	const tabKeys = computed(() => {
+		return items.value.reduce<string[]>((acc, item) => {
+			if (item.tab) {
+				acc.push(item.tab)
+			}
+			return acc
+		}, [])
+	})
+	const localModelValue = ref<string | undefined>()
+	const activeTabKey = computed({
+		get: () => {
+			return (
+				props.modelValue || localModelValue.value || tabKeys.value?.[0]
+			)
+		},
+		set: (newValue) => {
+			localModelValue.value = newValue
+			emit('update:modelValue', newValue)
+		},
+	})
+	const hasNavModifiers = computed(() => {
+		return [
+			'tabs',
+			...(Array.isArray(props.navModifiers)
+				? props.navModifiers
+				: props.navModifiers?.split(' ') ?? []),
+		]
+	})
+	const onNavClick = (item: NavItemTab) => {
+		if (item.tab) {
+			activeTabKey.value = item.tab
+		}
+	}
+	const navItems = computed(() =>
+		props.items.map((item) => ({
+			current: item.tab === activeTabKey.value,
+			...item,
+		})),
+	)
 
 	// bem css classes
 	const bemCssClasses = useModifiers('vv-tab', modifiers)
-
-	const localItems = computed(() => {
-		return items.value.map((item, index) => {
-			return {
-				...item,
-				id: item.id || `tab-item_${index}`,
-			}
-		})
-	})
-
-	/**
-	 * Triggers when the item is clicked.
-	 * @private
-	 * @event click
-	 * @param [NavItem, string] item - the clicked item
-	 */
-	function onClick(item: NavItem) {
-		if (!item.disabled) {
-			emit('click', item)
-			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-			activeTab.value = item.id!
-		}
-	}
 </script>
 
 <template>
 	<div :class="bemCssClasses">
-		<VvNav :items="localItems" modifiers="tabs full" @click="onClick" />
-		<!-- #region tab content -->
-		<article
-			v-for="item in localItems"
-			:id="item.id"
-			:key="item.id"
-			:class="{ target: activeTab === item.id }"
-			class="vv-tab__panel"
-		>
-			<slot :name="`${item.id}`" />
-		</article>
-		<!-- #endregion tab content -->
+		<!-- #region nav -->
+		<VvNav
+			v-bind="{
+				items: navItems,
+				modifiers: hasNavModifiers,
+			}"
+			@click="onNavClick"
+		/>
+		<!-- #endregion -->
+
+		<!-- #region panels -->
+		<template v-for="(item, index) in items" :key="index">
+			<article
+				v-if="item.tab"
+				:class="{ target: activeTabKey === item.tab }"
+				class="vv-tab__panel"
+			>
+				<slot :name="`panel::${item.tab}`" />
+			</article>
+		</template>
+		<!-- #endregion -->
 	</div>
 </template>