+++ src/components/VvTab/index.ts
@@ -1,13 +1,19 @@
 import { ModifiersProps } from '@/props'
-import type { NavItem } from '../VvNav'
+import type { NavItemTab } from '@/types/nav'
 
 export const VvTabProps = {
 	...ModifiersProps,
+	navModifiers: {
+		type: [String, Array] as PropType<string | string[]>,
+	},
+	modelValue: {
+		type: String,
+		default: '',
+	},
 	items: {
-		type: Array as PropType<NavItem[]>,
-		required: true,
+		type: Array as PropType<NavItemTab[]>,
 		default: () => [],
 	},
 }
 
-export const VvTabEvents = ['click']
+export const VvTabEvents = ['update:modelValue']